프론트 폴더 분리 및, cors 추가, host 변경
