+++ front/public/manifest.json
