+++ front/src/component/CookingList.jsx
@@ -20,7 +20,7 @@ const CookingList = () => {
       event.target.classList.add('CookingList__tabItem--active')
     } 
 
-    axios.get(`${host.server}/recipe/list?req=${event ? event.target.attributes.nation.value : `KOR`}`).then((result) => {
+    axios.get(`${host.server}/recipe/list?req=${event ? event.target.attributes.nation.value : `KOR`}`, {withCredentials:true}).then((result) => {
       setCookingList(result.data)
     }).catch( error => { console.log('failed', error) });
   }, []);
