+++ front/src/component/InventoryItem.jsx
