+++ front/src/host.js
@@ -1,0 +1,3 @@
+module.exports = {
+    server: (process.env.NODE_ENV==='production') ? null : "http://localhost:3001",
+};
