+++ front/src/prevCode/Stock.js
