+++ front/src/prevCode/asides/cuisines/JapanCuisine.js
