+++ front/src/prevCode/asides/cuisines/WesternCuisine.js
