+++ front/src/prevCode/kichens/JapanKichen.js
