--- src/host.js
@@ -1,3 +1,0 @@
-module.exports = {
-    server: "http://localhost:3001",
-};
